refactor: fractured the engine into several classes
